692 - more password translations
